+++ delivery-service/src/main/java/com/oingmaryho/business/delivery_service/application/service/DeliveryService.java
@@ -1,4 +1,4 @@
-package com.oingmaryho.business.delivery_service.application;
+package com.oingmaryho.business.delivery_service.application.service;
 
 import com.oingmaryho.business.delivery_service.application.dto.mapper.DeliveryApplicationMapper;
 import com.oingmaryho.business.delivery_service.application.dto.request.*;
@@ -17,37 +17,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class DeliveryAdminService {
+public class DeliveryService {
+
     private final DeliveryRepository deliveryRepository;
-
-
-    @Transactional
-    public DeliveryCreationResponseServiceDto createDelivery(Long userId,
-                                                             UserRoleType userRole,
-                                                             DeliveryCreationRequestServiceDto requestServiceDto) {
-
-        // TODO 1. 배송 경로 요청 (허브 도메인에 요청)
-        // GET -> List<HubResponseDto> hubRoutes
-
-        // TODO 2. 배송 담당자 생성 (유저 도메인에 요청) , 소속 업체 id 조회 (업체 도메인에 요청)
-        // GET -> List<UserResponseDto> users
-        // 허브 배송 담당자: 전체 10명, 업체 배송 담당자: 각 허브당 10명 존재
-
-        // stream().map -> List<DeliveryRoute> routes 생성
-        //      순차 배정 방식으로 허브 배송 담당자를 각 배송 경로에 매핑
-        //      정렬되어 온다면, index 값을 sequence에 매핑
-        // DeliveryManager 생성
-        //      허브 배송 담당자, 업체 배송 담당자의 경우, 매핑된 허브 경로의 출발 허브를 hubId로 지정
-        //      업체 배송 담당자의 경우, 소속 업체를 companyId로 지정
-
-        // TODO 3. 배송 생성
-        // Delivery delivery = DeliveryApplicationMapper.INSTANCE.toDelivery(managerId, hubRoutes[0].getDepartureId(),hubRoutes[hubRoutes.size()-1].getDestinationHubId(),requestServiceDto, routes);
-        // deliveryRepository.save(delivery);
-
-        // TODO 4. 배송 UUID 반환
-        // 메시지큐로 구현한다면, 주문 도메인에 UUID 메시지 전송
-        return null;
-    }
 
     @Transactional
     public DeliveryUpdateResponseServiceDto updateDelivery(Long userId,
@@ -55,8 +27,8 @@ public class DeliveryAdminService {
                                                            DeliveryUpdateRequestServiceDto requestServiceDto) {
         Delivery delivery = deliveryRepository.findById(requestServiceDto.id())
                 .orElseThrow(() -> new EntityNotFoundException("delivery not found"));  // TODO 커스텀 예외 처리
-
         // TODO 권한 확인
+
         delivery.update(requestServiceDto);
         return DeliveryApplicationMapper.INSTANCE.toUpdateResponseServiceDto(delivery.getId());
     }
@@ -67,7 +39,6 @@ public class DeliveryAdminService {
                                                                        DeliveryUpdateStatusRequestServiceDto requestServiceDto) {
         Delivery delivery = deliveryRepository.findById(requestServiceDto.id())
                 .orElseThrow(() -> new EntityNotFoundException("delivery not found"));  // TODO 커스텀 예외 처리
-
         // TODO 권한 확인
         delivery.updateStatus(requestServiceDto);
         return DeliveryApplicationMapper.INSTANCE.toUpdateStatusResponseServiceDto(delivery.getId());
@@ -77,6 +48,7 @@ public class DeliveryAdminService {
     public void deleteDelivery(Long userId,
                                UserRoleType userRole,
                                DeliveryDeletionRequestServiceDto requestServiceDto) {
+
         // TODO 권한 확인
 
         Delivery delivery = deliveryRepository.findById(requestServiceDto.id())
@@ -86,10 +58,11 @@ public class DeliveryAdminService {
 
     }
 
-    @Transactional(readOnly = true)
+    @Transactional(readOnly =true)
     public DeliveryResponseServiceDto GetDeliveryDetail(Long userId,
                                                         UserRoleType userRole,
                                                         DeliveryDetailRequestServiceDto requestServiceDto) {
+
         Delivery delivery = deliveryRepository.findById(requestServiceDto.id())
                 .orElseThrow(() -> new EntityNotFoundException("delivery not found"));  // TODO 커스텀 예외 처리
 
@@ -113,14 +86,14 @@ public class DeliveryAdminService {
         return deliveries.map(DeliveryApplicationMapper.INSTANCE::toDeliveryResponseServiceDto);
     }
 
-    @Transactional(readOnly = true)
+    @Transactional(readOnly =true)
     public DeliveryRouteResponseServiceDto GetDeliveryRouteDetail(Long userId,
                                                                   UserRoleType userRole,
                                                                   DeliveryRouteDetailRequestServiceDto requestServiceDto) {
         DeliveryRoute route = deliveryRepository.findByRouteId(requestServiceDto.id())
-                .orElseThrow(() -> new EntityNotFoundException("delivery route not found"));
+                .orElseThrow(() -> new EntityNotFoundException("delivery route not found"));   // TODO 커스텀 예외 처리
 
-        return DeliveryApplicationMapper.INSTANCE.toRouteResponseServiceDto(route);   // TODO 커스텀 예외 처리
+        return DeliveryApplicationMapper.INSTANCE.toRouteResponseServiceDto(route);
     }
 
     @Transactional(readOnly = true)
@@ -139,5 +112,6 @@ public class DeliveryAdminService {
                 requestServiceDto.customPageable());
 
         return routes.map(DeliveryApplicationMapper.INSTANCE::toRouteResponseServiceDto);
+
     }
 }
